Update ldflags for static-pie compatibility
In go 1.15, the default linkmode is `internal` when specifying
`-buildmode=pie`. Previously, the default linkmode was `external` for this
buildmode, so it needs to be specified manually now, as static-pie mode does
not appear possible (yet) with the internal linker.

Note that this change is backwards compatible with go 1.13 in case we need to
downgrade later.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>